uploadservice/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
greenlet>=3.0.0
aiomysql==0.2.0
pymysql>=1.1.0
cryptography>=41.0.0

# Excel processing
pandas==2.2.0
openpyxl==3.1.2

# Message Broker
aio-pika==9.3.1

# HTTP Client
httpx==0.26.0

# Testing (opcional)
pytest==7.4.4
pytest-asyncio==0.23.3